[dogstatsd] make flush interval configurable
The default flush interval for dogstatsd is 10 seconds. This is a
reasonable value for web page backends, but may be too coarse for
applications where each node serves hundreds of requests per second.

To support finer granularity, add a new configuration key
'statsd_flush_interval' and default it to 10. Because lowering this
value affects how often statsd calculates the distribution data points,
add a comment to explain the effect and discourage lowering the value
unless really necessary.